python/requirements.txt
flask
flask_sqlalchemy
python-decouple
tenacity
gladiator
sqlalchemy_utils